% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansTest.R
\name{meansTest}
\alias{meansTest}
\title{Significance test for difference between 2 mean values}
\usage{
meansTest(av1, sd1, nd1, av2, sd2, nd2)
}
\arguments{
\item{av1}{mean value of group 1}

\item{sd1}{standard deviation of group 1}

\item{nd1}{number of group 1}

\item{av2}{mean value of group 2}

\item{sd2}{standard deviation of group 2}

\item{nd2}{number of group 2}
}
\value{
dav 1:deviation of mean values

cohen_d 2:Cohen's d

t 3:t-value

pv 4:P-value P(T<t)

cl_l 5:95\% CI lower limit

cl_u 6:95\% CI upper limit

txj 7:flag(1 or 0) for significance/insignificance
}
\description{
\code{meansTest} statistical test for mean values
}
\examples{
# meansTest(60, 10, 20, 55, 15, 25)
}
