% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansTest.R
\name{meansTest}
\alias{meansTest}
\title{Significance test for difference between 2 mean values}
\usage{
meansTest(av1, sd1, nd1, av2, sd2, nd2)
}
\arguments{
\item{av1}{mean value of group 1}

\item{sd1}{standard deviation of group 1}

\item{nd1}{number of group 1}

\item{av2}{mean value of group 2}

\item{sd2}{standard deviation of group 2}

\item{nd2}{number of group 2}
}
\value{
dav deviation

cohen_d cohen's-d

t t-value

pv P-value

cl_l CI_lower_limit

cl_u CI_upper_limit

txj significant_or_not
}
\description{
\code{meansTest} statistical test for mean values
}
\examples{
# meansTest(60, 10, 20, 55, 15, 25)
}
